Split XTCE parser from packet_config
