+++ data/config/screen.yaml
@@ -127,7 +127,7 @@ SUBSETTING:
       values: .*
 NAMED_WIDGET:
   summary: Name a widget to allow access to it via the get_named_widget method
-  description: To programatically accesss parts of a telemetry screen you need
+  description: To programatically access parts of a telemetry screen you need
     to name the widget. This is useful when creating screens with buttons that
     read values from other widgets.
   warning: get_named_widget returns the widget itself and thus must be operated
